+++ services/madoc-ts/__tests__/capture-models/editor/revision-store.test.ts
@@ -47,15 +47,15 @@ describe('Revision store', () => {
       expect(store.getState().currentRevisionId).toEqual('7c26cf57-5950-4849-b533-11e0ee4afa4b');
 
       expect(store.getState().currentRevision).toMatchInlineSnapshot(`
-        Object {
+        {
           "captureModelId": "b329e009-1c8a-4bed-bfde-c2a587a22f97",
-          "document": Object {
+          "document": {
             "description": "",
             "id": "3353dc03-9f35-49e7-9b81-4090fa533c64",
             "label": "Simple document",
-            "properties": Object {
-              "name": Array [
-                Object {
+            "properties": {
+              "name": [
+                {
                   "description": "The name of the thing",
                   "id": "eafb62d7-71b7-47bd-b887-def8655d8d2a",
                   "label": "Name",
@@ -67,8 +67,8 @@ describe('Revision store', () => {
             },
             "type": "entity",
           },
-          "revision": Object {
-            "fields": Array [
+          "revision": {
+            "fields": [
               "name",
             ],
             "id": "7c26cf57-5950-4849-b533-11e0ee4afa4b",
@@ -170,23 +170,23 @@ describe('Revision store', () => {
       selectRevision({ revisionId: '18e0d089-e87c-448f-9ef2-3282a44cf6d2' });
 
       expect(store.getState().selector.selectorPaths).toMatchInlineSnapshot(`
-        Object {
-          "007a5ace-b1a4-49ca-8dd3-c78aee2d5409": Array [
-            Array [
+        {
+          "007a5ace-b1a4-49ca-8dd3-c78aee2d5409": [
+            [
               "person",
               "56405dc7-b910-45e0-8ade-898594276795",
             ],
-            Array [
+            [
               "firstName",
               "9a55a096-4a79-46b0-8111-d9775d074a14",
             ],
           ],
-          "05a99c13-eedd-4a14-940a-f774fc461ca4": Array [
-            Array [
+          "05a99c13-eedd-4a14-940a-f774fc461ca4": [
+            [
               "person",
               "cc2570c7-b3a5-4e5b-8f3d-3b3693769969",
             ],
-            Array [
+            [
               "firstName",
               "c3434721-19f7-4402-9bfb-1e11875a0de0",
             ],
@@ -308,13 +308,13 @@ describe('Revision store', () => {
       const selector = store.getState().selector.selectorPaths[selectorId];
 
       expect(selector[0]).toMatchInlineSnapshot(`
-        Array [
+        [
           "paragraph",
           "159621fb-4f93-4cd7-a394-5a1141fc1091",
         ]
       `);
       expect(selector[1]).toMatchInlineSnapshot(`
-        Array [
+        [
           "lines",
           "64e82cb7-16f8-432e-b2b7-3828233a134c",
         ]
@@ -369,16 +369,16 @@ describe('Revision store', () => {
           },
         },
         `
-        Object {
+        {
           "allowMultiple": true,
           "description": "Describe a person",
           "id": Any<String>,
           "immutable": false,
           "label": "Person",
           "labelledBy": "firstName",
-          "properties": Object {
-            "firstName": Array [
-              Object {
+          "properties": {
+            "firstName": [
+              {
                 "id": Any<String>,
                 "label": "First name",
                 "revision": "0e29f176-aeeb-4bf3-a92c-d64654e29c90",
@@ -386,8 +386,8 @@ describe('Revision store', () => {
                 "value": "",
               },
             ],
-            "lastName": Array [
-              Object {
+            "lastName": [
+              {
                 "id": Any<String>,
                 "label": "Last name",
                 "revision": "0e29f176-aeeb-4bf3-a92c-d64654e29c90",
@@ -432,7 +432,7 @@ describe('Revision store', () => {
           selector: { id: expect.any(String) },
         },
         `
-        Object {
+        {
           "allowMultiple": true,
           "description": "All of the lines inside of a paragraph",
           "id": Any<String>,
@@ -440,9 +440,9 @@ describe('Revision store', () => {
           "label": "Line",
           "labelledBy": "text",
           "pluralLabel": "Lines",
-          "properties": Object {
-            "text": Array [
-              Object {
+          "properties": {
+            "text": [
+              {
                 "allowMultiple": true,
                 "description": "Single word, phrase or the whole line",
                 "id": Any<String>,
@@ -450,7 +450,7 @@ describe('Revision store', () => {
                 "multiline": false,
                 "pluralField": "Text of lines",
                 "revision": "c8bb939a-7a76-4b15-9f77-81375519128c",
-                "selector": Object {
+                "selector": {
                   "id": Any<String>,
                   "state": null,
                   "type": "box-selector",
@@ -461,7 +461,7 @@ describe('Revision store', () => {
             ],
           },
           "revision": "c8bb939a-7a76-4b15-9f77-81375519128c",
-          "selector": Object {
+          "selector": {
             "id": Any<String>,
             "state": null,
             "type": "box-selector",
@@ -612,18 +612,18 @@ describe('Revision store', () => {
         expect(revisionSelectState.revisionSelectedFieldProperty).toEqual('firstName');
         expect(revisionSelectState.revisionSelectedFieldInstance).toEqual('9a55a096-4a79-46b0-8111-d9775d074a14');
         expect(revisionSelectState.revisionSubtreeField).toMatchInlineSnapshot(`
-                  Object {
-                    "id": "9a55a096-4a79-46b0-8111-d9775d074a14",
-                    "label": "First name",
-                    "selector": Object {
-                      "id": "007a5ace-b1a4-49ca-8dd3-c78aee2d5409",
-                      "state": null,
-                      "type": "box-selector",
-                    },
-                    "type": "text-field",
-                    "value": "second first name",
-                  }
-              `);
+          {
+            "id": "9a55a096-4a79-46b0-8111-d9775d074a14",
+            "label": "First name",
+            "selector": {
+              "id": "007a5ace-b1a4-49ca-8dd3-c78aee2d5409",
+              "state": null,
+              "type": "box-selector",
+            },
+            "type": "text-field",
+            "value": "second first name",
+          }
+        `);
       });
       it('should remove the selected field when navigating', () => {
         revisionPopSubtree(undefined);
@@ -792,7 +792,7 @@ describe('Revision store', () => {
           },
         },
         `
-        Object {
+        {
           "allowMultiple": true,
           "description": "Single word, phrase or the whole line",
           "id": Any<String>,
@@ -800,7 +800,7 @@ describe('Revision store', () => {
           "multiline": false,
           "pluralField": "Text of lines",
           "revision": Any<String>,
-          "selector": Object {
+          "selector": {
             "id": Any<String>,
             "state": null,
             "type": "box-selector",
@@ -845,10 +845,10 @@ describe('Revision store', () => {
           revisedBy: expect.any(Array),
         },
         `
-        Object {
+        {
           "id": Any<String>,
           "revisedBy": Any<Array>,
-          "state": Object {
+          "state": {
             "height": 40,
             "width": 30,
             "x": 10,
@@ -885,9 +885,9 @@ describe('Revision store', () => {
           id: expect.any(String),
         },
         `
-        Object {
+        {
           "id": Any<String>,
-          "state": Object {
+          "state": {
             "height": 4,
             "width": 3,
             "x": 1,