+++ services/madoc-ts/__tests__/capture-models/helpers/filter-empty-structures.test.ts
@@ -34,40 +34,40 @@ describe('filterEmptyStructures', () => {
     const structure = filterEmptyStructures(model);
 
     expect(structure).toMatchInlineSnapshot(`
-      Object {
+      {
         "description": "Contains single choice with single text field",
         "id": "cdee3083-5f74-43c1-9d80-d97349f5fa69",
-        "items": Array [
-          Object {
-            "fields": Array [
+        "items": [
+          {
+            "fields": [
               "label",
             ],
             "id": "f0c7f99c-9e4b-4211-8dc5-f94c1287690b",
             "label": "Completely valid",
             "type": "model",
           },
-          Object {
-            "fields": Array [
+          {
+            "fields": [
               "label",
             ],
             "id": "c9db4483-9421-4370-abab-2c24ee656de2",
             "label": "Partially valid",
             "type": "model",
           },
-          Object {
+          {
             "description": "This is a nested choice",
             "id": "983b5776-4a2a-4107-a50d-eab94f0d2c0e",
-            "items": Array [
-              Object {
-                "fields": Array [
+            "items": [
+              {
+                "fields": [
                   "label",
                 ],
                 "id": "c40a1d4b-6bcc-4dd0-89f3-f8146a0837d0",
                 "label": "Nested - partially valid",
                 "type": "model",
               },
-              Object {
-                "fields": Array [
+              {
+                "fields": [
                   "label",
                 ],
                 "id": "6961fa68-0886-4bbe-b680-9f6004a2fe12",
@@ -78,12 +78,12 @@ describe('filterEmptyStructures', () => {
             "label": "Nested choice - partially valid",
             "type": "choice",
           },
-          Object {
+          {
             "description": "The second choice in the list",
-            "fields": Array [
-              Array [
+            "fields": [
+              [
                 "person",
-                Array [
+                [
                   "firstName",
                   "lastName",
                 ],
@@ -93,12 +93,12 @@ describe('filterEmptyStructures', () => {
             "label": "Valid Person",
             "type": "model",
           },
-          Object {
+          {
             "description": "The second choice in the list",
-            "fields": Array [
-              Array [
+            "fields": [
+              [
                 "person",
-                Array [
+                [
                   "firstName",
                 ],
               ],
@@ -107,16 +107,16 @@ describe('filterEmptyStructures', () => {
             "label": "Partially valid person",
             "type": "model",
           },
-          Object {
+          {
             "description": "This is a nested choice",
             "id": "0c1675fc-96d0-42f3-bfb9-f9f8fa56ffa0",
-            "items": Array [
-              Object {
+            "items": [
+              {
                 "description": "The second choice in the list",
-                "fields": Array [
-                  Array [
+                "fields": [
+                  [
                     "person",
-                    Array [
+                    [
                       "firstName",
                       "lastName",
                     ],
@@ -126,12 +126,12 @@ describe('filterEmptyStructures', () => {
                 "label": "Valid Person",
                 "type": "model",
               },
-              Object {
+              {
                 "description": "The second choice in the list",
-                "fields": Array [
-                  Array [
+                "fields": [
+                  [
                     "person",
-                    Array [
+                    [
                       "firstName",
                     ],
                   ],