+++ services/madoc-ts/src/frontend/shared/atlas/RegionHighlight.tsx
@@ -36,6 +36,7 @@ export function RegionHighlight({
   const saveCallback = useCallback(
     (bounds: any) => {
       console.trace();
+      console.log(region, bounds)
       if (isEditing) {
         onSave({ id: region.id, x: region.x, y: region.y, height: region.height, width: region.width, ...bounds });
       }
@@ -62,6 +63,7 @@ export function RegionHighlight({
           e.stopPropagation();
           onClick(region);
         }}
+        relativeStyle
         onMouseEnter={() => {
           onHover(region);
         }}