+++ services/madoc-ts/src/frontend/shared/capture-models/_components/ViewDocument/components/ViewField.tsx
@@ -7,11 +7,20 @@ import { BaseField } from '../../../types/field-types';
 import { FieldSection } from '../ViewDocument.styles';
 import { ViewSelector } from './ViewSelector';
 
-export function ViewField({ field, fluidImage }: { field: BaseField; fluidImage?: boolean }) {
+export function ViewField({
+  field,
+  fluidImage,
+  revisionId,
+}: {
+  field: BaseField;
+  fluidImage?: boolean;
+  revisionId?: string;
+}) {
   const ref = useRef<HTMLDivElement>(null);
 
   const helper = useSelectorHelper();
-  const selector = field.selector ? resolveSelector(field.selector) : undefined;
+
+  const selector = field.selector ? resolveSelector(field.selector, revisionId) : undefined;
   const selectorId = selector?.id;
   const [isOn, trigger] = useDecayState();
 
@@ -24,7 +33,7 @@ export function ViewField({ field, fluidImage }: { field: BaseField; fluidImage?
         }
       });
     }
-  }, [helper, selectorId]);
+  }, [helper, selectorId, trigger]);
 
   useEffect(() => {
     if (selectorId) {
@@ -35,9 +44,9 @@ export function ViewField({ field, fluidImage }: { field: BaseField; fluidImage?
         }
       });
     }
-  }, [helper, selectorId]);
+  }, [helper, selectorId, trigger]);
 
-  if (!field.selector) {
+  if (!selector) {
     return <FieldPreview key={field.id} field={field} />;
   }
 
@@ -49,7 +58,7 @@ export function ViewField({ field, fluidImage }: { field: BaseField; fluidImage?
       onMouseLeave={() => (selectorId ? helper.clearHighlight(selectorId) : null)}
       data-highlighted={isOn}
     >
-      <ViewSelector selector={field.selector} fluidImage={fluidImage} />
+      <ViewSelector selector={selector} fluidImage={fluidImage} />
       <FieldPreview field={field} />
     </FieldSection>
   );