+++ services/madoc-ts/src/frontend/shared/capture-models/helpers/field-post-filters.ts
@@ -1,6 +1,9 @@
 import { CaptureModel } from '../types/capture-model';
 import { BaseField } from '../types/field-types';
-import { isEntityList } from './is-entity';
+import { isEntityEmpty } from '../utility/is-entity-empty';
+import { isEmptyFieldList } from '../utility/is-field-list-empty';
+import { isEntity, isEntityList } from './is-entity';
+import { traverseDocument } from './traverse-document';
 
 type PostFilter = (fields: BaseField[]) => BaseField[];
 
@@ -9,7 +12,31 @@ export function filterEmptyFields(fields: BaseField[], parent?: CaptureModel['do
     return fields;
   }
 
-  if (isEntityList(fields)) {
+  if (isEntityList(fields) && fields.length > 1) {
+    // 1st is this allowMultiple?
+    const first = fields.find(f => !f.revision) || fields[0];
+
+    // In this case, we have a list of entites - but the templated entity might be in there.
+    // This is a rough "isEntityEmpty"
+    if (first && first.allowMultiple) {
+      const findEmpty = fields.filter(r => !r.revision);
+      if (findEmpty.length) {
+        const toRemove: any[] = [];
+        for (const found of findEmpty) {
+          if (isEntity(found)) {
+            if (isEntityEmpty(found)) {
+              toRemove.push(found);
+            }
+          } else {
+            if (isEmptyFieldList([found])) {
+              toRemove.push(found);
+            }
+          }
+        }
+        return fields.filter(inst => toRemove.indexOf(inst) === -1);
+      }
+    }
+
     return fields;
   }