+++ services/madoc-ts/src/frontend/shared/capture-models/utility/is-field-list-empty.ts
@@ -1,3 +1,4 @@
+import { getEmptyBorder } from '../editor/input-types/BorderField/BorderField';
 import { BaseField } from '../types/field-types';
 
 export const isEmptyFieldList = (fields: BaseField[]) => {
@@ -6,7 +7,17 @@ export const isEmptyFieldList = (fields: BaseField[]) => {
   }
   for (const field of fields) {
     if (field.value) {
+      // Hack. We need an "isEmpty" on the field definitions I think.
+      if (field.type === 'border-field') {
+        return field.value.size === 0;
+      }
       return false;
+    }
+
+    if (field.selector) {
+      if (field.selector.revisedBy && field.selector.revisedBy.length) {
+        return false;
+      }
     }
   }
   return true;